move broadcasting code to separate file
